pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.39.0",
    "duckdb>=1.3.0",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.49.0",
    "xlsxwriter>=3.2.9",
]
